KerasCaptcha/requirements.txt
tensorflow
keras
pillow
captcha
numpy